Converted the Windows app to use PyInstaller instead of py2exe. Pyinstaller is available from pip: pip install PyInstaller
Also, I fixed the Inno Setup script to include a proper SnapPy icon
for the taskbar.

This currently builds a working app. (I used the free Visual Studio
C++ for python 2.7 to compile everything except CyPari.)  There are
two issues that I know of:

1. The app will crash with the 4.0.3 version of iPython.  There is
a bug in the traitlets module which was fixed with commit
721abf5ce13... at https://github.com/ipython/traitlets.  So get the
zip file (no git in my Msys) and reinstall traitlets from there.
It is a pure python module, so you can also edit line 110 of
traitlets.py by hand (TypeError -> Exception).

2. The Inno Setup uninstaller does not remove the snappy database
files.  The Windows uninstall reports success and removes SnapPy from
the registry, but the SnapPy directory will still exist with the
sqlite files in it.  This needs to be removed by hand after running
the uninstall.