EPIC COMMIT MESSAGE FOR A TWO LINE CHANGE:
The kernel functions "triangulations.c:data_to_triangulation" and
"terse_triangulation.c:terse_to_tri", which are used by the methods
"_from_string()" and "_from_bytes" of Triangulation respectively, call
"orient", "create_cusps", "create_edge_classes", and
"orient_edge_classes" in different orders. The historical motivation
for this is unclear, but in any event it creates the following
issue.

For repeatability, if one saves a Triangulation T to a file and loads
it again, the default presentation for its fundamental group must be
unchanged. If it so happens that T is an orientable manifold that was
created by a call to "data_to_triangulation" where the data in
question was *not* consistently oriented, then, prior to this commit,
the presentation for the fundamental group often changed. Such data
arises frequently from isosigs and the like, and "terse_to_tri"
doesn't have this problem.

The issue was that the presentation for the fundamental group depends on
the edge orientations (since those give the relators) and this is
affected by the order in which one calls "orient" and
"create/orient_edge_classes" when the initial data is not itself
consistently oriented.

This commit makes it so that "data_to_triangulation" calls
"create_edge_classes", and "orient_edge_classes" a second time when it
has oriented the triangulation manually, as opposed to being told in
advance that the data represents an orientable manifold.