This commit removes "fake enums" from SnapPea.h and kernel_typedefs.h.
The "fake enums" are constructions like this:
typedef int Something;
enum {
   zero = 0,
   one = 1
}
They were used in the kernel because iteration over values of an enum
is not allowed in C++. The actual effect is equivalent to a typedef int
with #define statements specifying names for the values.  There are two
reasons not to use these.  First, they prevent the compiler from
enforcing correct use of the enum.  Second, they obfuscate the code and
confuse automatic documentation generators.