Python version to place vertices of base tetrahedron: when trying to match the way the SnapPea kernel placed them, consider all even permutations of the tetrahedron instead of brute force trying to find matches among all potential places where the vertices could be.
In theory, the SnapPea kernel could trick the interval implementation in doing the wrong matches. This might be paranoid, but I think the philosophy should be that we can prove the interval result being correct no matter what the SnapPea kernel gave us.

Also using only one version of are_close that works for both intervals and Sage/SnapPy numbers.