Snap now computes the matrices by explicitly ungluing the manifold to obtain the fundamental polyhedron.
In theory, a simplex A could have a vertex at infinity, but the old method to visit tetrahedra could encounter a simplex B where no vertex is at infinity on way to visit simplex A so that the vertex at infinity needs to be computed instead of copied. The new method combinatorially identified all vertices of the simplices in the fundamental polyhedron belonging to the same vertex of the fundamental polyhedron in a vertex class and assigns infinity only once.

Further refactoring.