Changing how the cusp topology is stored so that we can support arbitrary vertex links.
Cusp::is_finite and Cusp::topology is replaced by an enum for orientability. The Euler characteristic is promoted.

Euler characteristic and orientability are initialized to values indicating that they have not been computed yet.

The CuspTopology enum is extended to cover all cases. get_cusp_topology and set_cusp_topology convert between the CuspTopology enum and the Euler characteristic and orientability. The former raises a coding error if the Euler characteristic or orientabiltiy have not been computed yet.

The function compute_cusp_orientability can compute the cusp orientability. peripheral_curves also still computes the cusp orientability - but only if the Euler characteristic is zero.

Note: previously, all vertex links were assumed to be spheres, Klein bottles or tori. The Euler characteristic was used to distinguish between the sphere and the other possible vertex links. The code to compute peripheral curves was also in charge of distinguishing between the Klein bottle and torus. The new code computes the cusp orientability when traversing the cusps.