Removing verbose_form (which was just added a couple of days ago) from FundamentalGroup.meridian and longitude.
We might change the type to represent words from bare strings to another class that has string like API for backwards compatibility but also knows about the context so that, e.g., you cannot try to concat two words in letters from different groups. It might be more appropriate to have methods to convert a word to the verbose and integer list form on that class.

Avoiding the proliferation of verbose_form and as_int_list_forms until then.