Improving fundamental polyhedron in python.
There is a loop trying different permutations of the initial tetrahedron to match the vertices to those of the SnapPy kernel.
In the old version, it was also trying different branches of the square root of the shape - but since shapes always have positive imaginary part, there should never be ambiguity when taking the square root.

The new version also makes sure that the permutation that makes the vertices match the vertices of the SnapPy kernel is unique.

This ensures even further that M.verify_hyperbolicity(holonomy=True) always returns the same PSL(2,C) matrices (just with different precisions). The function rather fails than being tricked into a different permutation for the initial tetrahedron if we changed the shapes stored in the SnapPea kernel.