Faster slicing obstructions
Frank Swenton discovered that a nontrivial amount of time was being
spent factoring the elementary divisors of the covers, and this could be
avoided since we only care about F_q part of the homology for q small.

Also, added some flexibility in specifying the pairs (p, q) to be
considered.