Length spectrum: fix for global spine radius.
It is still unclear whether the old bound was correct or not. I have compared the length spectrum results for the first 5 geodesics with and without the optimization checking against the global spine radius for at least half of the census manifolds and found no difference except in the words used to represent geodesics sporadically (which was never guaranteed to be stable to begin with).