Improve canonization method (#130)
Modifies the kernel function `proto_canonize` so that it only returns proto-canonical triangulations with no flat tetrahedra, where here "flat" means a shape with very small positive or negative imaginary part.

This is needed because the kernel function `canonical_retriangulation` returns the wrong answer when given a proto-canonical triangulation with flat tetrahedra.