Fixes a subtle orientation issues with Dehn-fillings of the Triangulation.triangulation_isosig (#131)
Fixes a subtle orientation issues with Dehn-fillings of the Triangulation.triangulation_isosig.

When specifying ignore_curves, we want to preserve the orientation of the core curve rather than of the Dehn-filling curve. This preserves the spun structure properly.

I discovered this as follows.

m007(3,1) aka Vol3 has no known geometric spun-triangulation. However, it has a unique three-fold cover. And that cover has a geometric spun-triangulation. I wanted a compact isosig for this spun-triangulation.

Here is a one isosig: 'gLLMQcbeefefpjaqupw_BBab(1,0)'

But I don't need the peripheral curves, so let's drop them.

>>> M = Manifold('gLLMQcbeefefpjaqupw_BBab(1,0)')
>>> isosig = M.triangulation_isosig(ignore_cusp_ordering=True,ignore_curves=True)
>>> isosig
'gLLMQcbeefefpjaqupw(-1,-1)'
>>> Manifold(isosig).solution_type()
'contains negatively oriented tetrahedra'

So somehow we got the opposite Dehn-filling orientation giving us the geometric solution:

>>> Manifold('gLLMQcbeefefpjaqupw(1,1)').solution_type()
'all tetrahedra positively oriented'

Here is the geometric spun-triangulation and its mirror image:

>>> M=Manifold("gLLMQcbeefefpjaqupw(1,1)")
>>> N=Manifold("gLLMQcbeefefujumapn(2,-1)")

But the result was different:

>>> M.triangulation_isosig(ignore_cusp_ordering=True,ignore_curves=True)
"gLLMQcbeefefpjaqupw(1,1)"
>>> N.triangulation_isosig(ignore_cusp_ordering=True,ignore_curves=True)
"gLLMQcbeefefpjaqupw(-1,-1)"

For the unoriented spun-triangulation, the direction of the core curve matters. The direction of the filling curve is determines by the orientation of the core curve and the manifold. In particular, we need to reverse the filling curve when reversing the manifold direction.

This is consistent with Manifold.reverse_orientation():

>>> M = Manifold("m004(3,4)")
>>> M.reverse_orientation()
>>> M
m004(3,-4)

That is, it reversed the meridian and the second Dehn-filling coefficient to effecitively reverse the Dehn-filling curve. Reversing the manifold orientation as well keeps the core curve oriented the same way.